Handle unsupported audio channels on Linux.
The SbPlayer implementation relies on that the Cobalt side doesn't send
any unsupported format to it.  This is usually true when running YT
player on Cobalt as the format is pre-filtered by isTypeSupported().

However during unittests it is possible that unsupported data (in this
case, 5.1 audio) is sent to the SbPlayer implementation, which causes
DCHECKs internally.

Add checks in PlayerComponents::Create() to ensure that the audio
parameters are supported on the platform.  Also guard member functions
of FilterBasedPlayerWorkerHandler on creation failures.

b/62056243

Change-Id: Ifa3d994daf863a6e9924a4e8fccc3f09949145bb

Former-commit-id: 17d1450989025d98459cea9f859c1adcf93e3ab2