Fix FromJSValue for Window object.
Do not assume that all wrappable objects in JavaScripts are proxy
objects.

Also fix WrapperFactory::IsWrapper() to use HasWrapperPrivate()
instead of relying on JS_GetPrivate(), which asserts instead of
retuning null when the object is not a wrapped object.

This fixes FromJSValue when used to convert a Window object from
Javascript to a native object.

b/62071511

Change-Id: Ide81fa619bb4641c81e922068e82c8f7f4125d6b

Former-commit-id: 8d7f93987936271aba4d206ad6475cd77104ca9c