Android: Improve video renderer seeking
Previously, when seeking in |VideoRenderer|, we were counting warmup
frames that were well before the time that we were seeking to against
ourselves as dropped frames, even though that they indeed should not
be rendered.  This change adds logic to check for that case.  As a
result of this, the stats for nerds dropped frames counter is much
more accurate, and because we can actually stall audio playback while
we are busy burning through our warmup frames, we do end up winning
a couple of frames back after a seek.

Change-Id: Iff729d4a9dcdd40c20901e262a476b77c164aa27
(cherry picked from commit 642236608f529d0fbd5be3acdc89f18ee2ec2351)

Former-commit-id: 748d64c3a04ff5cf0c45a9b0789b0b991c76e956