Organize webdriver_benchmark tests
Previously, all of the webdriver_benchmark tests were located in a
single directory. In preparation of new tests, this has been changed
so that the performance tests are located within a "performance"
directory, which is further divided into "non_video" and "video"
tests.

The top level performance test collections are still located in the
"tests" directory, but have had redundant parts of their name removed.

This will make it easier to follow the tests contained within test
collections when more types are added, such as "memory" or "endurance".

Change-Id: I33a104347a93ca2317a13e44edd108bc4dc0594d

Former-commit-id: b6875c7d735fd551c5231b2ee884f166faedd478