Support HTMLMediaElement::suspend event.
The HTMLMediaElement should fire suspend event when there isn't any
network activities in progress.  The spec is fairly vague on this but is
clear that after a progressive video finishes playback, a suspend event
should be fired as it is expected that there is no more data
downloading.

Now FetcherBufferedDataSource will notify WebMediaPlayerImpl when
network activity starts/stops, which in turn set the network state of the
HTMLMediaElement accordingly.

b/62142728

Change-Id: I74440f9505ca6b8746e52b609ba11c68d6e9db2f

Former-commit-id: ae2f3f0a4ad6c44014ed4dee05ee2a01a8b831b9