Hookup mouse events from Android to Starboard.
This converts mouse button and motion events to the corresponding
Starboard Input Events.

For platforms where the Starboard API is older than
SB_POINTER_INPUT_API_VERSION, wheel events are reported as key
presses, and chorded button interactions are not reported as such due
to limitations in the API.  See go/cobalt-mouse-support for details.

b/37275032

Change-Id: Ic8c33e4a4328b6fba4978d2e13ec3c1482e0a47d

Former-commit-id: 29bd0d1c6850c1b8b246112a39e9931ffa40215f