Android: Cut down on playback object allocations
I noticed that we were getting some really bad GC pause times during
playback (16ms - 60ms on the upper end), large enough to negatively
impact playback.  These changes are the result of tracking where the
majority of allocations are happening (both in count and size), and
optimizing for minimizing for amount of intentionally dropped frames,
which are tied to application frame rate.  When playing a 4k VP9
60fps video on Nvidia Shield, forcing a 50hz refresh rate, this
took us from 25% intentionally dropped frames to roughly ~16.7%,
which is roughly the theoretical limit.

Change-Id: Id1eaeffab68e35d16c26c7655a295ddc429e59c8

Former-commit-id: c5562f161264d2b483a8addd5582ddffd2147e13