Fix bug in SbSocketWaiter
Previously, |vector.erase()| was being called on a vector, which caused
elements after the one being deleted to be shifted down.  This caused a
bug, since there were other data structures that stored indexes to
refer to sockets.  This CL changes the erase operation in a way so that
none of the indexes are invalidated.

Fixes b/62225194.

Change-Id: Id8b44b35fe3a8cb76dcc8f75f4cef833069d7473

Former-commit-id: cbfaa8f35a216950d466f902d1a7c51c1d930328