GYP: Allows variables in include paths
Expand environment & command line variables in include path.
This fixes a major problem with platform specific gypi includes
which are platform-specific paths. These dynamic paths could be
determined from variables, however gyp does not support these.

This turns out to be a major problem in gyp that many devs are
complaining about.

For cobalt this manifests as a huge issue for
deployment scripts, which are platform specific but can't be
included at the executable target section. The work around has
been to use gyp target dependency linking, but this has the issue
that variables cannot be passed using this method. As a result,
it is impossible for the deployment script to get the target
executable as a variable.

To put it another way:

STATUS QUO:

nb_test ('executable_name': 'nb_test')
  -> 'dependencies' (allows variable expansion)
    -> variables like 'executable_name' is chopped off!!!
      -> <(starboard_path)/platform_deploy.gyp:platform_deploy',
        -> includes <HARD CODED PATH TO GYPI FILE>

This change allows dynamic include paths, which allows a platform
specific gypi file to be included at the site of the executable
target.

NOW ALLOWED

nb_test ('executable_name': 'nb_test')
  -> includes <(starboard_path)/platform_deploy.gypi
    -> has access to 'executable_name'

This issue affects android and now xb1.

The fix was found on the chromium website at the following:
https://bugs.chromium.org/p/gyp/issues/attachmentText?aid=1005

Note that the only variables that can be used in include paths are
ones that are defined on the command line and environment variables.

Change-Id: I94e09fdc5ad7d2c67667a254e79a1b9e6921a8e0

Former-commit-id: 7446badd05e746536a468cf27b4caaeee5494f36