Bump up default pexpect timeout for raspi run_test
Previously the default pexpect timeout in the raspi's run_test, which
is used by the raspi_launcher, was set to 120 seconds. However, the
layout_benchmarks can take longer than that to output anything on the
raspi-0, which causes the test to fail. 300 seconds should be plenty
of time for any case and should fix flakiness with the benchmark tests
on the raspi-0.

Additionally, pexpect's expect timeout has been dropped to 60 seconds.
This one isn't nearly as variable.

Change-Id: I569e1b85fb4f7dc04e3e868f620f13c4d7fb6b51

Former-commit-id: 9225da0443ebb3d31ac7a5609939d217bf6b796c