Support aria-describedby property in build-in screen reader
The aria-labelledby attribute is similar to aria-describedby in that
both reference other elements to calculate a text alternative, but a
label should be concise, where a description is intented to provide more
verbose information.

Reversed engineering the ChromeVox behavior when both aria-labelledby and
aria-describedby are presented, aria-describedby is ignored in this
case.

b/62643181

Change-Id: I1b46d064cc2dd26d52ae591b4bbcf9acab63ab41

Former-commit-id: 6065f91e88e99a3bc2eae58eee0650fc25ad0911