Track launch times in webdriver benchmark startup script
Previously, the webdriver benchmarks initially started on about:blank,
would measure the time for that to fully load, and then would load
***REMOVED*** several times, measuring counts and durations. This didn't
provide the most accurate measure of the total startup time, since
some initialization happens the first time ***REMOVED*** is launched, that
won't happen on subsequent loads. Typically, the time measured is
~500ms lower on the RasPi.

The logic has now been changed so that the webdriver benchmarks
initially load ***REMOVED***. All of the durations are tracked from the
launch until ***REMOVED*** is fully usable. This provides the most accurate
measurement of startup time. Then the reloads occur as before. These
provide a larger sample size and are more useful for tracking changes
in performance.

Change-Id: I2511f326d62c096a443d8d92973ecf771bcc75c3

Former-commit-id: 9c128c5f470c72e2fffd75c78aab7ca6c7a2b4e1