Always fallback to custom GLES to render YUV images.
Depending on the result of some GLES checks done upon Skia
initialization, Skia may use GL_RED texture formats or GL_ALPHA
texture formats for single channel textures, but there's no
way to query its decision.

We used to be creating textures outside of Skia and then passing
them in, but if we get this format mismatched, then we end up having
a bad day.  Since we already have code now to render YUV images
outside of Skia, just use that all the time so that we don't have
to deal with this problem.

Change-Id: I6ad6184dcbde9e509f70346d2d9e0a73f6f42d69

Former-commit-id: 22670db3f6dbd29a81edb91f24548c77eb7d8a55