Disable caching of JS constructor code when jitting disabled
Cobalt recently started definining JS_CODEGEN_{ARCH}, because
SpiderMonkey relies on it for architecture specific files that are
always needed. To continue disabling jitting, Cobalt relies on
COBALT_DISABLE_JIT to disable jit-related code.

Mozjs-45 uses JS_CODEGEN_NONE to determine whether or not to cache the
constructor code of UnboxedPlainObjects, so this change caused the
caching to become enabled.

This caused two issues:
1. Performance is slightly slower in Cobalt when using this code.
2. RasPis crash in this code.

Due to these issues, the logic has been modified so that we no longer
cache constructor code for UnboxedPlainObjects when jitting is
disabled, which should fix the RasPi crash and result in slightly
better performance.

Change-Id: Ifaf51e6227cf95031756f4544464e82365447b4b

Former-commit-id: 5885d9f2aa98060a31504562bf70a1e864c4e783