Drop |kPrerollFrames| from 8 to 1 for video playback
Local testing did not show any stuttering on the raspberry pi when
dropping the preroll frames from 8 to 1 and it seems to result in a
better user experience.

In terms of performance on the RasPi2, this change reduces the median
video playback delay from 2.75sec to 2.29sec, reduces the mean delay
from 2.85sec to 2.45sec, reduces the min delay from 1.29sec to .98sec,
and reduces the 95th percentile from 4.15sec to 3.5sec.

The idea for this came from nathanrogers, so all credit should go to
him.

Change-Id: I04a8effbc1860cb082890f47bfef230a9bfeb1e1

Former-commit-id: 6d33c48c46ccb29fb4a785d0baec02004d7d2f45