Re-add support for falling back to custom GLES to render YUV images.
This is undoing the revert for:
https://lbshell-internal-review.git.corp.google.com/#/c/65672/

which broke on the directgles renderer platforms.

The additional changes to fix directgles include recognizing when
we are rendering to surfaces whose origins are in the top left instead
of bottom left.

Surfaces that are intended to be used as textures later will have their
origin set to the top left, whereas surfaces intended to be used as
display render targets will have their origins set to the bottom left.
While Skia knows these settings (and uses them to orient their draw
calls), they are not queryable publicly from a SkCanvas object, so
we do our own [ad-hoc] tracking of surface origin.

Change-Id: Ic9d4c057cea2a220cdf205cb18598ac6221cdbfb

Former-commit-id: 52a90840c9f1f47ba6734b72a64c30108c451cc1