Add StartupTimer to cobalt/base
Previously, there was no single arbiter of truth on the start time for
the app. Application set a value in its constructor, while WebModule
used a different value that was loader initialized. As a result of
this difference, the WebModule value showed a time that was about 10
ms earlier than Application time.

To standardize the time and make it globally accessible, the logic in
WebModule has been moved into base/StartupTimer and both Application
and WebModule have been updated to use it.

Change-Id: I105ad87058385a5b5642bd902f551800c6e46466

Former-commit-id: 29e16f3951a028f260169d3114494179b4a31429