Make webdriver benchmarks less flaky
BrowseToWatchTest has been modified so that it'll allow for up to 10%
of videos failing to play. If this amount is exceeded, it'll raise an
exception.

The number of videos that are played during the test has been bumped
up to 50 to further reduce noise.

Additionally, images occasionally took longer to load than the old
processing timeout, so the value has been doubled.

Change-Id: Ib5117a1e6fcb1f79136132355a6f28041f5f070d

Former-commit-id: e3bc9b76edd6c65833d758863f026b2169be594d