Improve BrowseToWatch benchmarks logic
BrowseToWatch now checks for if an ad is playing before a video. If it
is, then it waits for the ad to become skippable, so that it won't keep
seeing ads, and culls the event from its results.

Cookies are now cleared when load_tv() is called, so that the videos
that are watched during one load_tv() run, won't impact the videos seen
during the next run.

Change-Id: Ib19079c2899acaeb9e7f0cb514421e86fec66c4b

Former-commit-id: 7b7b664f5291213469fcd9e10575cc678e9f9a49