Fix race condition where video output mode may not get updated.
This could result in the layout tree thinking that the video
output mode is invalid indefinitely, resulting in a black
rectangle being rendered for the video instead of the video
content.

While it was more likely for this to happen when playing
encrypted content, it can also be reproduced for standard
content with a strategically placed sleep call.

b/63144152

Change-Id: I65de52aae8424643343b57b73f7f889a73f1e463

Former-commit-id: cb16293bbae451667721c5ec32230d8fc00a91df