Initialize the skia HardwareBackendImage ASAP to reduce frame spikes.
On the raspi-2, it costs about 8-9ms on average to create and upload
the texture data for each thumbnail. When scrolling vertically, one
frame ends up initializing 4 thumbnails, then the remaining animation
frames had 0 thumbnails to initialize. This smooths out that cost since
it takes 20+ ms to decode an image, so the rasterizer now incurs one
upload cost every 20+ ms.

Change-Id: Ie0d8770a88c383c614137c4455403d035f6e7047

Former-commit-id: 63b214a1cbd2f243c805cfbda342f9b3fb5c6c1b