Always use stacking context when closer than containing block
Previously, in the case where a positioned box had a z-index of 0, the
containing block would always be used as its stacking context. However,
this caused issues in the cases where the containing block was a more
distant ancestor than the nearest stacking context. In these cases, the
draw order could be wrong on the child (b/62982044) or opacity might not
be applied (b/62982509).

The logic has been changed so that a containing block more distant than
the nearest stacking context is never selected as the stacking context
for a box. The only case where a containing block will be chosen is when
it is closer than the nearest stacking context (meaning that the box
will still be contained within the correct stacking context). Choosing
the containing block in these cases is needed to prevent bugs with
overflow hidden.

However, even with these changes, overflow hidden bugs can still occur
when a box has a non-zero z-index value. In this case, the nearest
stacking context must be used, but a nearer containing block may be
skipped over, which can cause overflow hidden values in the containing
block to not be obeyed.

This logic resolves b/62982044 and b/62982509.

Change-Id: I62b3dbb20457a07547940fab09d69d774d3815d1

Former-commit-id: 5b2310fbde5bcaa4fae3cf9829af27d4e5df2eb5