Model ResourceProviderStub font metrics on Roboto
ResourceProviderStub has been modified so that it provides font metrics
that match our version of Roboto. Additionally, the width and glyph
bounds calculations within it are now modeled to mimic the value
provided by the '0' character in Roboto. This enables Cobalt to generate
fairly accurate layouts with ResourceProviderStub when running ***REMOVED***.

Change-Id: Idf40ca0ba79af6ffaa625a59a0312a7c1667817d

Former-commit-id: 9935dd89dc79c668404ff6030da8682b4e6e1d53