Fix animation bug with display: none.
Keep track of whether a node has an ancestor whose display is
set to none, so that we can know if that node is actually
displayed or not.

Also refactor duplicated code in updating the computed style for
pseudo elements and the main element.

Update Window::GetComputedStyle() to call
UpdateComputedStyleOnElementAndAncestor() instead of the
more expensive UpdateComputedStyles().

b/63252142

Change-Id: I261316db0f14b4c52dd1b862806c3231e11588cb

Former-commit-id: 5da86a15ddc563c7df04e3c06e6e562c038b2797