Fix bug in GetOffsetFromStackingContextToContainingBlock()
Previously, there was a bug in
GetOffsetFromStackingContextToContainingBlock() where the content box
offset from the padding box was not being added to |relative_position|
for absolutely positioned boxes in the case where containing block
was an ancestor of the stacking context. This has been fixed and a
layout test has been added for the case.

Additionally, the code has been cleaned up and optimized by tracking
the relationship between the containing block and stacking context
(which is determined when the child box is added to the stacking
context). Doing this has allowed the second function
GetOffsetFromStackingContextToContainingBlock() to be combined with the
first function and for a happy path to be created when the stacking
context and the containing block are the same, which does not do any of
the extra work.

This resolves b/63714503.

Change-Id: I917a76fea96b1792b2a1dba27c56c3e8b192d882

Former-commit-id: a191399f3074276b96949b36ba90dd803edbdb70