De-couple CValCollectionEntryStats OnFlushCallback from Visibility
Previously, any class using OnFlushCallback from
CValCollectionEntryStats or CValCollectionTimerStats needed to use a
Visibility of CValPublic. Switching it to CValDebug would require the
class to use ifdefs for its callback logic since FlushResults wouldn't
be defined in Gold.

The logic has been changed so that FlushResults is now available to
CValCollectionEntryStatsStub, which means that CValCollectionEntryStats
and CValCollectionTimerStats with an OnFlushCallback defined can now
use a Visibility of CValDebug without breaking the build or requiring
ifdefs.

Change-Id: I2bac15c4b21a2bd6d4ab020b55fc574a553e7eaa

Former-commit-id: bef0fadcf668345a0fcc1f8f4bc0186b11903471