Call SetMultithreadProtected on D3D11DeviceContext
We need to set multithreaded mode on D3D11 device's immediate context
in order to also use it for the hardware decoder.

Also, we need to set this on the renderer thread or before rendering
starts, otherwise we're accessing the context from multiple threads w/o
enabling thread protection.

There's no existing Starboard layer opportunity to do this, so
we'll just set it here at create time.

Change-Id: I35f7f55e82bfc9c0588189708278d88b97818a5c

Former-commit-id: 3c34c9e97c8a6e1af6b563bcdebe1bf9680b3ca8