Add h5vcc.sso
This CL adds h5vcc.sso interface, which allows Javascript developers to
get api keys, and secrets to do Oauth.  Each starboard port is
responsible for implementing the interface in sso/sso_interface.h, and
implementing a static function that will return an instance of a class
that implements the interface.

Change-Id: I5e7794cafc164c81e5f9c595e399350c3db3b343

Former-commit-id: 704c3e05b5baca2d1dcbc7879608e2f521b642c7