Change DLOG to DCHECK when calculating offsets
Previously, we had a DLOG for the case where a transform existed
between the stacking context and containing block. However, this
shouldn't be possible because a transform forms both a stacking
context and containing block. Encountering this case indicates a
pretty serious bug in the layout logic, so I'm switching it to
DCHECK so that it'll crash in devel.

Change-Id: Idb85f71f34b1081fa2d1129b658dd7fb8bd09058

Former-commit-id: ae76e1d1ba81428a905b1556b1078d544d237ddc