Standardize containing blocks usable as stacking contexts
Previously, non-positioned stacking contexts and relatively positioned
elements would never consider a nearer containing block for their
stacking context. The reason for this was that their actual containing
block was potentially an in-flow, non-positioned element, which would
not participate in the stacking context itself.

However, this caused issues because their stacking context could end
higher up the tree than an absolute containing block. In the case
where a later sibling was absolutely positioned, the stacking order
could end up wrong, because the relatively positioned element would
end up in a higher stacking context and be added after the shared
parent, meaning that it'd now be drawn after the absolute positioned
element.

To solve this, all non-fixed positioned elements and stacking contexts
are added to the nearest ancestor between the absolute containing block
and the stacking context. This guarantees that all non-fixed position
siblings with a z-index of 0 will be added to the same stacking context
and will maintain the proper draw order.

There is still a potential issue with fixed position elements having
draw order bugs; however, this isn't a newly introduced issue. To
resolve this, we'll need to decouple overflow hidden from the stacking
context (in which case, we'll be able to always use the stacking context
as the actual stacking context for all elements).

This fixes b/63864375.

Change-Id: Ie8b25e0936c2a1522a6b16b58828dee2a52bc763

Former-commit-id: f2087555019941fb5a7f59f5dae9719b8f31244e