Fix suspend/resume when animated webp is playing back.
Fix issue where Cobalt would crash when suspending while
an animated WebP is playing back.

This was caused because a reference to the AnimatedWebPImage
object would be kept around longer than expected because it
was referenced from a lingering render tree.  This is fixed
by having the render tree reference a simpler "FrameProvider"
frame container class that can be divorced from the
AnimatedWebPImage.

Note that this also (implicitly) fixes an issue where the
renderer would need to acquire a lock that is also acquired
during image frame decodes, meaning that the renderer may
occasionally stutter on that lock if a decode is taking
place.

b/63628978

Fix issue with animated webp images not restarting playback
after resuming.

This was caused by the AnimatedImageTracker maintaining a
stale reference to an old animated webp image from before the
suspend occurred, and it would then have a buggy perception of
image playback state.  This is fixed by resetting the tracker
on suspend.

b/63540954

Change-Id: I6adeb2be20ffdd6abb33e8de4b6c9eaf80430d1d

Former-commit-id: 8946a31edca204ee0ef18a6ea44414eef1700972