Change NOTREACHED to DLOG in offset calculation logic
It is a valid case for the NOTREACHED to be hit here. It can occur when
the stacking context is above the containing block and is a statically
positioned element with opacity. A descendant box with a non-zero
z-index would be able to hit this case.

I plan to have it fixed shortly, but I want this in to avoid an
unnecessary crash in the meantime.

Change-Id: I979525705030841eb0576e872c35bf781dedff6e

Former-commit-id: 134512cffde740fdae01883ceb4b7e96d9b96931