Improve containing block offset calculation logic
Previously, there were a couple failure cases in
GetOffsetFromStackingContextToContainingBlock(), where it would
incorrectly calculate the offset from the containing block to the root
element rather than the offset to the stacking context:
1. Where the stacking context was above the containing block and was
   not positioned.
2. Where the stacking context was above the containing block and a
   fixed position containing block was between them.

The logic has been changed so that if the root element in the tree is
encountered, the walk is reversed so that it walks from the stacking
context to the root and subtracts the encountered offset. This removes
the extra, incorrect offset that was being added during the first walk.

Layout tests have been added for these cases.

Change-Id: Idb0839ab1f4743d858699c39085cf3c3ca520ef5

Former-commit-id: d10e3600f074c28a82306bd5b6a1525aec634f14