xb1: Allow NV12 EGL_D3D_TEXTURE_ANGLE
Stage one of allowing NV12 EGL_D3D_TEXTURE_ANGLE textures.

In order to do decode to texture, we must wrap a D3D texture
that we get from the video decoder in a GL texture.

We use the pre-existing EGL_D3D_TEXTURE_ANGLE mechanism, but
we need to:

- Disable the (unnecessary) assumption that this texture can be
  drawn to (can be a render target).

- Allow the NV12 color space (existing validation functions disallow it)

- When setting up for drawing (as a D3 Sharer Resource View),
  make sure to configure the appropriate color channel.

Currently this change produces only one texture, binding the Y channel
to the R channel. A follow up will allow the UV to be bound to other
channels.

Change-Id: Id89328b6574b5fcffc9cc7aeea207bb7786d37b1

Former-commit-id: 5a3f4f1bfa84a08e76efac65a22fccb41dcb20c2