Always apply overflow hidden when appropriate
Previously, when a box was added to a stacking context child container
higher up in the tree than its containing block, any overflow hidden
from containing blocks that came between the box and the child container
in the layout tree was not applied. This has been fixed. Now, overflow
hidden is always applied in these cases. The use of z-index no longer
impacts whether or not overflow hidden is applied.

Layout tests have been added for this case.

Change-Id: Ia382136ad394a12cab9d7c7346d27092ed1ae5ac

Former-commit-id: 0192ec0286b796b6e6da6a65ce9166312e1c1a9c