xb1: Magic trick for NV12 textures
DirectX NV12 textures must be bound either to an "R8" output (for
luma) or "R8G8" output (for chroma) when their "Shader resource
views" are created.

In order to select this when creating a GL texture for
this DirectX texture, we make use of the DirectX "SetPrivateData"
mechanism.

When making a GL texture that's meant to bind to the chroma/R8G8 outpu,
we set a special "private data" value DirectX texture. We then check
that while making the GL texture and store it. We then make a brand
new GL texture from the same DirectX texture, but this time with
the private data field modified.

The private data has a GUID of {3C3A43AB-C69B-46C9-AA8D-B0CFFCD4596D}

Change-Id: I8d6fc395bb25914b6bf7d390f8d99cefc355eae0

Former-commit-id: 897337e3b6a17b565ffd0121aa3b637573014b85