Handle non-object this value in bindings
Our bindings code will attempt to convert the this argument when
calling a bindings getter/setter/method into an object.  Thanks to
the wonderful flexibility of JavaScript, this is a not a safe
assumption.  We must first check if the this arg we received is
actually an object or not, and throw in the case that it isn't

This fix is not required for mozjs-24, as the JSAPI interface for the
analogous functions will guarantee that the this arg we receive is an
object via a different signature.

b/64128577

Change-Id: I6ffaca707f22a53bbaeb33fc5aa048a354a7c926

Former-commit-id: 9d3923b582cdc2264a0657a7cd33f419fd24a4e2