xb1: Ignore unused warnings in qa/gold
Often variables are only used in DCHECKs, which causes warnings to
appear when DCHECKs are compiled out in qa or gold builds.

This CL ignores these warnings in qa and gold builds.  The warning will
still appear in debug/devel builds if there is a variable that is
genuinely unused.

Change-Id: I532709aa0621486b9a264844c91dc6d674864a45

Former-commit-id: cf4a515c947243b7509bdc75b6844fe722c88a70