Don't clear decoder buffer cache during resuming.
On low end devices it is possible that the audio is played at a faster
pace then the appending of the video data.  During resuming it is
possible that video data is cleared before it was played.  This can be
solved by not clearning decoder buffer cache during resuming.

This fix is taken from https://cobalt-review.googlesource.com/3470.

Change-Id: I650eb7abaed9cb808f453993e7ce956ac24ed056

Former-commit-id: afd3f3056f786299f3eff6ccd9cd042b62c7bb8d