Remove incorrect DCHECK in XMLHttpRequest
It is possible for an XMLHttpRequest to send itself in one of its
event handlers, which means multiple references created by the
Wrappable XMLHttpRequest to it will exist (one in send, and one in
the current handler).  Thus, it is incorrect to have a DCHECK that
ensures only one reference has been created from the Wrappable
XMLHttpRequest.

b/64267518

Change-Id: I888ec5fc0df10284a2b05d2fcfd4ab68174a6f1f

Former-commit-id: 5894e50b8ac1ac9c7cba24785bc81e253e704fcf