Fix infinite loop bug in ImageDataDecoder::DecodeChunk()
Previously, ImageDataDecoder::DecodeChunk() could get into an infinite
loop if it was decoding an image where it was unable to make progress
but |state_| was never being set to kError.

This was being triggered by JpegImageDecoder failing to read the header
and the issue being identified as not all of the header data being
available. When this occurred, but the full 32768 bytes were in the
buffer, ImageDataDecoder::DecodeChunk() would go into an infinite loop
making calls to DecodeChunkInternal().

This resolves the more dangerous of the two bugs in b/37539419.

Change-Id: Ibd08b80abc265e0788fc36c35ab49565680fc849

Former-commit-id: 06521e015a4d4ae45bc7ab18a62785db040396a4