Increase max buffer size in ImageDataDecoder
There was previously a bug in ImageDataDecoder::DecodeChunk() where, if
the first chunk was larger than the 32768 max buffer size, the buffer
could grow indefinitely. This was because buffer size is subtracted from
the max buffer size to determine the available bytes; when buffer size
was the larger value, the unsigned result wrapped to become a massive
positive value.

Ironically, this bug allowed the large background images in Unplugged to
usually be decoded; they require around 300KB to read the header, so by
having an unbound max buffer size they were decodable. However, in the
case where the first chunk was not larger than 32768 bytes, then the max
size was obeyed, and the image could not be decoded.

To resolve this, I'm upping the max buffer size to 4MB. I view this as
more of a sanity check value than anything else. Additionally, we're
no longer allowing the data buffer to grow larger than the max size.

This fixes b/37539419.

Change-Id: I6ff51f4e1a4251871da653d1f7d84279c1a0e008

Former-commit-id: b78e4a785a6e91635deb46f2527a1459ffd13b41