Use padding box as containing block with 'fixed' position boxes
Previously, there was a mysterious bug where even after the containing
block to stacking context offset calculation logic was fixed to properly
handle cases that included transforms, both of the disabled layout tests
that tested this functionality were producing incorrect results.

It turns out that the underlying bug was with fixed position elements
not using the padding box of the transformed element as their containing
block, which is how all major browsers handle this.

It is a bit of an odd case, because the only guidance the spec provides
is that the viewport always forms the containing block. However, in
practice, elements with transforms form the containing block of fixed
elements, and, when these have padding, that padding should be included
in the containing block calculation.

This also eliminates the scary "Containing block offset calculations
that include transforms may not be positioned correctly" warning that
was showing up when running ***REMOVED***.

The two disabled layout tests dealing with stacking contexts and
containing blocks have been enabled, and a new test explicitly testing
fixed position boxes and transforms with padding has been added.

Change-Id: I85dde4971d32eb14ba335265c4aa948fdafc9708

Former-commit-id: c1e46da713428c1aa97b7292c2c3effcc3c9605d