Tighten cases where overflow hidden invalidates cross references
The logic invalidating cross references as a result of children being
moved to a split sibling from a container with overflow hidden has
been tightened so that the cross references are only invalidated when
stacking context children exist.

Additionally, the code has been cleaned up and the comments improved so
that it's easier to follow.

Change-Id: Ia696d84fc334ad8cb7c8299369bb3a9b74cfe81b

Former-commit-id: 87142d750299ad647641e361d047fd5379ca77aa