Fork thread_create.cc on Android to add custom onThreadFinish logic.
Android platforms should call DetachCurrentThread() before being
shutdown.  This was formerly being done in the pthread local
value key destructor, but the implementation of DetachCurrentThread()
may itself have been relying on pthread local value code, so
this is not the safest place to call that function.  The logic is
now moved to instead call DetachCurrentThread() before signaling
to pthreads that the thread is completed.

b/64352137

Change-Id: I19f4112a0e0da91daeda1981ffb6da77b6770b16

Former-commit-id: dca8653fb74332749beca02ceaaca3c95d6c6eaf