Wait for processing complete before checking for survey
Previously, it was possible for the survey to not have been cleared
from the previous load before it was checked again on the next load.
Now, the processing complete check has been moved to before the survey
shelf to ensure that everything is fully loaded when the survey is
checked.

Change-Id: I14c3b105f9652e5d219faa16b10a029b3c1689fa

Former-commit-id: c37733b8d9eb5639478deca189ed30e4f0f2bc3d