NV12 SbDecodeTarget's can specify texture format
Previously, any NV12 texture had to have it's luma plane assigned
to the alpha plane and it's chroma plane assigned to the blue-alpha
planes.

However, some platforms (Microsoft Direct X) don't work that way.

Allow the NV12 texture format to be specified to also allow
luma on the red and chroma on the red-green, as DirectX required.

Change-Id: Iccaa0d75aaf705a52a388c2fb0a268bb14ea4c7a

Former-commit-id: 96a567fad00316e4184c1434f1b973270631038f