Increase the web module stack size
The PS3 was hitting the stack size limit and having JS execution fail
with the previous sizes. It needed around 575KB on Devel and 505KB on
on QA to avoid hitting the stack limit while running the RTK Realtek
"many" playlist.

I'm upping the stack size to 832KB in Devel and 768KB in QA to give us
a bit of breathing room with the stack sizes. This will resolve
b/64683861 when it is cherry-picked to COBALT_9, but will require a
different fix on COBALT_7.

Change-Id: Ic9918397d0cd4813f0f681cc12714af69d21b1a4

Former-commit-id: 1285b5985c6ea4a45b9037e476bcdabc10b4eb82