Add support for 'any' type in dictionary
Limited support has been added for using the 'any' type in
dictionaries. This is needed for both ErrorEventInit and
CustomEvent.

There are a few things to note with this:
1. The only default value that is supported with 'any' is null. The use
   of any other default value will generate a build error.
2. The dictionary keeps ScriptValue::StrongReference to the 'any'
   value. The reason for this is that dictionary does not have a
   wrappable to use with ScriptValue::Reference.
3. FromJSValue() does not work with script values in dictionaries.
4. The |has.*_| variable is only added to members that do not have a
   default value. This is due to the variable being superfluous for
   members with a default value.

Tests have been added for the 'any' type being in dictionaries.

NOTE: Thanks to ***REMOVED***@ for a good portion of this code.

Change-Id: Ie9e95d4d9e4aa6b8f3331b44df6cb2a8bd346c81

Former-commit-id: f9dbf5b37de85ebd10a3771cd26c8c85bd9adaf4