Fix invocations of find_private_files.py in GN
In GYP, external scripts in the form of ("<!@(python blah blah)") are
run with their current directory as the directory of the .gyp file
invoking them. However, in GN, external scripts are always run with
the current directory as the root build directory
(e.g. out/stub_debug). Fix the GN code to rebase paths back and forth.

Change-Id: I20a5c7313d863ab77ec7654187339cc9bfc112a8

Former-commit-id: 0993c5809f4ca2268b5484b0326cc7dcb7c72548