Make it so that <|(...) gyp variables don't pollute source directory.
Originally, <|(...) use in gyp would result in an output file being
written into the source directory.  This can mean that if the
contents of this file would change depending on your output gyp
config, you would always need to re-gyp before switching build
configs.  This change makes it so that the output file goes into
out/BUILD_CONFIG/gypfiles/... instead.

This is a partial cherry-pick of
https://codereview.chromium.org/38313004, with some modifications.

b/37672769

Change-Id: I922e62e04e68bd0105ea374cbd0641400dad14f0

Former-commit-id: 47cc4d07c2c105ff1d232c1ed727c662735a5ba4