Improve FindTopmostEventTarget() logic
Previously, the FindTopmostEventTarget() logic relied upon any
transforms occurring at (0,0) or not having any children to work
properly. In cases where this was not true, then the coordinate
being checked was incorrect.

This has been fixed. Now, the base offset of the transform is treated
as the new root for any of its children, and their offsets are
calculated in reference to it. All of the Get.*Box.*Edge() functions
have been modified so that they now have the option of going up to the
root element or stopping at the first transform encountered.

Additionally, the previous logic added a coordinate offset for each
absolute positioned box encountered. While this is correct for most
boxes, it adds extra offsets for fixed position boxes, which likely
won't count absolute positioned boxes as containing boxes.

The logic has now been modified so that this offset is automatically
applied in the Get.*Box.*Edge() functions, rather than requiring it to
be applied as an extra step. This also fixes the functions in
LayoutBoxes, which were including the containing box padding in their
calculations for absolutely positioned boxes.

Also, split sibling boxes are now checked in
TopmostEventTarget::ConsiderBoxes(); previously they were skipped.

Lastly, to track the implementation of layout tests for pointer events,
b/64904826 has been created.

Change-Id: I3d821652c98feb730ef6a97abf3885a8009f4189

Former-commit-id: 05df1ff489d6de278ceda9c93bd5b540486bc49b