Use Touchpad and stick range -1..1 with (-1,-1) at top left.
This changes the range for the touchpad to -1..1 with (-1,-1) at the
top left.

This changes the thumbstick to use negative values for the up and
left direction, matching existing implementation's axis directions.

The -1..1 is the 'unit distance' for a thumb controlled touchpad.
Since the PS4 controller is twice as wide as it is high, the X axis
range is -2..2.

b/64954544

Change-Id: I0cb236b9c8f25085ae5db4783dc25f3bc8af1336

Former-commit-id: e72fd336688ab758e2ce5a3e524fee2345287ab6