Fix an issue when libwebp decode stride does not match buffer stride.
Our libwebp decoder was erroneously expecting the stride set
during initialization to be honored by libwebp's decoder, but
libwebp was just ignoring that and re-writing its own stride.

This change just listens and adapts to libwebp's chosen stride
instead of expecting it to conform to the stride we give it.

This fixes an issue where some animated webp images can be
sheered because of a stride mismatch.  This was visible on
linux-x64directfb.

Change-Id: Idbfedc6f096d2cd441f412a50ca1e9f01866e7d4

Former-commit-id: bd52e8f920120f85009ac6d891289067d185e6e0