Make the NetworkModule IO thread high priority.
The NetworkModule IO thread pumps and forwards socket
traffic, and does not do any heavy CPU lifting.  Therefore,
it should be a high priority thread so that it can respond
to socket events ASAP.  This may be the source of PS3
endurance crashes when animated webp is active, as the
NetworkModule thread starves and so the kernel socket memory
starts filling up.

b/64612391

Change-Id: I2e8b9a9df5a84bc762f5822a382840bfbbdfd347

Former-commit-id: 82d157dec16f96fac6a122b6f7a2ede2d94e10df