Check inline style for map-to-mesh if computed style is unavailable.
It seems as though unplugged is creating media elements and setting
their src property *before* attaching the media element to the
document, and therefore when playback starts and we are trying
to figure out whether to use decode-to-texture (e.g. for 360 videos)
or punch through, we query whether the map-to-mesh CSS filter is
applied to the media element or not.  To do this check, we calculate
the computed style and then check it, but if the media element
is not attached to the document, it has an undefined computed
style, so in that case, we check the inline style instead.

b/64821185

Change-Id: I0ace73dbaf4ad2d6d10e88112f73cadc3a8f30a6

Former-commit-id: f68c583fdb16a2409be16e38a136ebe3337e866f