Fix preload race condition
Previously, there was a race condition with preload where the
offset_height that Cobalt provided ***REMOVED*** could be 0. This would cause
***REMOVED*** to set the global font-size to 0%, which would cause most of the
UI to become invisible.

This race condition was caused by StartOrResumeInternal() calling
UpdateFromSystemWindow(), which posts a SetSize task. This in turn posts
a resize event task. This resize event causes ***REMOVED*** to run the logic
which sets the global font-size. StartOrResumeInternal() also runs the
LifecycleObserver callbacks, which post an event to call
WebModule::Start(). For everything to work, the WebModule::Start() task
had to be posted before the resize event task. When this was the case,
the application was in the correct state when offset_height() was
called. If the resize event was posted first, then the application was
in the preloading state, and the layout needed by offset_height() was
blocked, resulting in 0 being returned for the offset_height.

To fix this, window now checks its visibility state before dispatching
a resize event. If it is visible, then it is immediately dispatched.
If it is not, then it sets a pending resize event, which causing the
resize event to fire when the visibility state changes to visible.

Additionally, the resize event is no longer posted, but is instead
immediately dispatched, making the order that it and WebModule::Start()
run deterministic.

This fixes b/64817452.

Change-Id: Ie2ee10b4094a0ccd6d91e40a4dc4563851593e68

Former-commit-id: 779c5b1775c6e7c108c83467b3057e9ed5120777