Improve Savegame logic
Previously, a significant number of superfluous saves were occurring
while running the app. For example, 10 saves would occur for someone
logged in on startup, 8 saves would occur for someone not logged in
at startup, and 10 saves would occur when clearing cookies.

The logic has been changed so that the last savegame data is cached,
and the new save does not occur if it matches the old save.
Additionally, the timers triggering a Flush() will be reset if an
immediate Flush() is triggered; before it would trigger an additional,
unnecessary Flush(). Lastly, the timer itself has been changed from
100ms to 500ms.

These changes reduce the number of saves at startup to 2 (from 10) and
the number of saves after clearing cookies to 1 (from 10).

StorageManager tests have been added for the new functionality.

This fixes b/64294032.

Change-Id: I5b2d9948f51db5e7f5bf0379fb852ed97110be4e

Former-commit-id: 76a2191ffed9e480320cc9e3ba3ef41784dcbd05