End playback when Pipeline is ended
Previously the HTMLMediaElement ends playback when the current time is
beyond the duration of the media.  Due to round up issues the current
time may be slightly off.  The Pipeline signals the end of playback to
WebMediaPlayerImpl using OnPipelineEnded(), now forward this to
HTMLMediaElement so it can end playback in such situation.

b/64211755
b/62103445

Change-Id: Iec45f27cf37156e73773bafdc5ab76295a0cf297

Former-commit-id: eb9f390e8c283fb43ac20f4b0c13f4702302d672