End playback when Pipeline is ended
Previously the HTMLMediaElement ends playback when the current time is
beyond the duration of the media.  Due to round up issues the current
time may be slightly off.  The Pipeline signals the end of playback to
WebMediaPlayerImpl using OnPipelineEnded(), now forward this to
HTMLMediaElement so it can end playback in such situation.

b/64211755
b/62103445

Change-Id: I6f650097b04570848a5c80d7ba614860f861a36d

Former-commit-id: 8b5ec1948669e1f24342d222f285035256960452