Prevent early garbage collection on nested XHR.
A new XHR request can be opened on the same XHR object in old request's  event listeners such as load. In that case, we want to prevent
calling garbage collection before all requests are fetched.
Right now IncrementActiveRequests() are called in send(). It might worth considering prevent calling AllowGarbageCollection if a new request called open() but not send().

b/65137671
Change-Id: I51a0863b67afdfb4f44519a813c01809d709fe0d

Former-commit-id: 724c26b190aea0dc7c77916c9210d98ce3736a72