Adding client for new starboard_app_launcher logic
This script allows users to run individual executable targets
in a cross-platform environment by using the new launcher logic.
For example, once the Android launcher is fully implemented, this
script could be used to run individual Android unit tests without
having to manually interact with ADB. Providing a target name to the
script, such as "audio_test", will allow the Android launcher to handle
all of the logic behind executing that target.

This script should not have to be modified as new launchers are
implemented.  However, if the structure of the abstraction changes
this script may need to change to reflect that.

Change-Id: I63de8853c376de86c113ee09ca8f232dc4c50d2f

Former-commit-id: a8d73185dcdb57fc9092d55fe0ed38c9a2a9d814