Add support for window.onerror
On script errors, Cobalt now generates an ErrorEvent, which it
dispatches synchronously from window. If the script has muted errors,
then no data is contained in the event; otherwise it is populated with
a message, filename, line number, column number, and error object
provided by mozjs.

If the error event was canceled, then the error is treated as being
handled and no logging is allowed. Otherwise, logging continues as
before.

This fixes b/62373983.

Change-Id: I41f50a3190d8e1fd82d2bee9ee65bbf038f6d9f7

Former-commit-id: 55587e829dca57c2298b20c7d353acf8917aa1d0