Updating Android exit code file logic.
Currently, when an Android app exits, Starboard
creates a file, then writes the exit code to that file and
closes it.  However, this means that anything searching for that
exit code file could find it before data has been written.  This
added logic writes the exit code to a temporary file and then renames
it to the intended name, meaning that the exit code file is not visible
in the filesystem until it has been written to.

Change-Id: I2548857cde05913468d2ea0dbfef7a7c04671ad2

Former-commit-id: 106036c317d3426798cdd89a476cb0f8c168f45b