Handle preload with disconnected network
Previously, if Cobalt started in preload with a disconnected network, it
would go into an infinite loop attempting to load invalid URLs. This was
because OnError() was called as a result of the network failure. This
attempted a retry URL; however, the H5vccURLHandler was not available
while in preload to handle the retry url, so OnError() simply triggered
another Navigate(), which triggered another OnError() call. During each
call, "h5vcc://network-failure?retry-url=" was prepended to the URL,
causing it to continuously grow larger.

The logic has been changed so that OnError() no longer calls Navigate().
It now starts a timer that calls OnErrorRetry(). The delay on this timer
uses an exponential backoff, depending on the number of times that
OnErrorRetry() has previously been called without a call of OnDone().
This backoff goes from 1 second up to a max of 1024 seconds (~17
minutes). Also, OnError() sets the OnError url. This URL is set until
Navigate() is called, at which time it is cleared. If start or resume
is called while the OnError url is set, then a navigate to the OnError
url occurs when starting the app.

Additionally, the system platform error logic has been moved from system
window into its own class, SystemPlatformErrorHandler, which is owned by
the browser (it did not use any window-specific logic). This allows the
browser to create the H5vccURLHandler immediately, rather than waiting
on a system window, which enables it to be available in preload and
guarantees that the platform will receive the same system network errors
regardless of the app state.

Lastly, NavigateInternal() has been combined with Navigate(). Now,
Navigate() contains all of the logic within NavigateInternal() but
calls itself if it isn't running on its message loop. However, it clears
|web_module_loaded_| regardless of the thread that it is cleared on.
This removes an unnecessary PostTask() in the case where Navigate() is
called on the browser module's thread.

This fixes b/64866015.

Change-Id: I819ebb0aac55df3fed80b8895f5532e3a3ba2139

Former-commit-id: c693e8b5299297b5d9f1e817049a310a84592540