Optimize direct-gles box shadow with rounded corners.
This splits up the geometry so that the relevant rounded corner is
passed through the vertex shader instead of forcing the fragment
shader to do the corner selection.

This also decouples DrawRectShadowBlur from DrawRectShadowSpread since
they no longer have any shaders in common.

This also tweaks the rounded rect clipping so it can accommodate large
radius values for the rounded corners.

Change-Id: Ie731ecad5f37bf294862eb3c2f1d6c717afa01e2

Former-commit-id: 8ad53382f85974f23ba4ba7708852b1462de1cea