Add JSGCThingTracer
This change adds the JSGCThingTracer tool to SpiderMonkey.
JSGCThingTracer is a tool that every 15 minutes, logs the gc things
that are currently alive in the JavaScript heap, as well as where and
when they were allocated.  Its primary use case is tracking down
JavaScript application side memory leaks, in which the hosted
JavaScript application requests an unbounded amount of memory over
time, that does get cleaned up on exit, and is correct to allocate
and hold on to from a Cobalt in isolation perspective.

b/65130215

Change-Id: I3c8f23969821715c50f5be590da1a53ef937870e

Former-commit-id: 2308775b9992118bb77a51482275a1213d02799d