xb1: Turn on HDCP when required
Implements SbMediaSetOutputProtection and SbMediaIsOutputProtected,
and implements the necessary logic in the PlayReady system to set
HDCP, as required.

It takes about 10 milliseconds on my setup (Xbox One S + Grundig TV)
(Xbox One S + Grundig TV).  A timer was added in case this ends up
being a source of avoidable latency in some setups.

The HdcpSession API doesn't really work as of UWP redstone version 2,
which is 10.0.15063.4082 .

b/65117856

Change-Id: Ib674eff825a3af61b6fbe6889b60de396703e32b

Former-commit-id: efe32d4b8f3d514ae5a3e7587247636bc1c83c83