Only re-render the scene if animations actually change.
Adjust animated bounding rectangle to only consider
regions that are actively animated.

This CL makes it so that our animation starting point
is the last animated render tree, and we then detect
if there are any changes before and after applying
animations.  If there are no changes, we bubble that
information up the tree by leaving it unchanged, so
that at a high level, we can quickly check if animations
were applied by testing if our "after animations" tree is
the same as the "before animations" tree.

Adjust fps counter to respond to a new periodic interval
fps timer as with animated webp enabled, the interval timer
is not very helpful (e.g. it will update with 1 sample
every 100ms).

You can see that this really frees up single processor
devices.  Here is a before/after trace from raspi-0:

Before: https://***REMOVED***
After: https://***REMOVED***

b/65130070

Change-Id: I3801d15852c8100640b4f08c1af50a1b84089e28

Former-commit-id: d7f62f89f29c49515cc4e819a564b80fd8a7d430