Style sheet removal triggers selector tree recreation
The selector tree is not currently designed to gracefully handle style
sheets being removed or the order of style sheets changing. In one case
it holds onto stale pointers and in the other case it maintains an
incorrect internal order for the rules.

To solve this, in the cases where a document has a style sheet removed
or its style sheets reordered, the selector tree is now destroyed and
recreated. This is likely to be a rare case and allows us to avoid
adding additional complexity to the selector tree.

Also, the matching rules invalidation logic has been changed so that
all descendants are traversed on invalidations. This eliminates the risk
of invalidations not occurring because an ancestor was already invalid.
However, only the originating element invalidates siblings now, because
in the case where the invalidation started higher up the tree, this is
already handled by the parent.

Rule matching tests have been added for these cases.

This fixes b/65257129.

Change-Id: I1d49d8d214338f58e3585faa64bc22437ff49fa3

Former-commit-id: 11890e2dc440ccdb1795ffe03a6257d94517168c