Check for non-invertible matrices with transforms
It is possible for transform matrices to be non-invertible, such
as when a scale of 0 is used. This was resulting in the coordinates in
ApplyTransformActionToCoordinates() being set to -NaN.

While this was not noticeable for us, because our test platforms always
fail NaN comparisons, ensuring that IsUnderCoordinate() never returned
true, this caused issues for LG. Our suspicion is that NaN was being
treated as 0 by LG, which caused boxes with the non-invertible
transforms to always be treated as being under the mouse after we
changed the border box logic in IsUnderCoordinate() to calculate the
offset from the transform rather than the viewport.

We now check for non-invertible matrices in
ApplyTransformActionToCoordinates() and return false if a non-invertible
matrix is encountered. In this case, the logic looking for boxes under
the coordinate in TopmostEventTarget earlies out and does not check any
contained children, as the coordinates will be invalid.

This likely resolves b/65227881, although we won't know for sure until
LG confirms the fix.

Change-Id: I713e138ba417726623391cc60f7d0a6ef8fba5b3

Former-commit-id: f23cea2281e17b908844cda49dde140409c5d572