Serialize transform scale to scaleX/scaleY.
This serializes transform values 'scale' to 'scaleX' or 'scaleY' in
some cases.  When the Y axis scale is 1, 'scaleX' is used.
Otherwise, when the X axis scale is 1, 'scaleY' is used.  In all
other cases, 'scale' is used.

This fixes a Karma test for the progress bar that expects this
behavior.

b/62307713 b/34893160

Change-Id: I22cb7dd68760e12fb92a87e4714ba095f6f11510

Former-commit-id: 5d195d031f609e34388611edd1761690863f516c