xb1: move eglGetPlatformDisplayEXT
Previously eglGetPlatformDisplayEXT was being called in main(),
but that caused secondary build targets (eg nb_test) not to link.

Unfortunately, we can't have the starboard target depend on ANGLE, since
ANGLE depends on starboard.

Instead, we can move this initialization to SbWindowCreate where it mostly
won't cause a problem.

Also, disable starboard_window_example for xb1 because it calls SbWindowCreate.

Change-Id: I685f475eebc91e95952a9e8b73bc4847306c3f22

Former-commit-id: 84046b770bde38fa1741475be393c043a7cef967