Delay Navigate() until after state update
Previously, the Navigate() during a Start() or Resume() happened before
the app state was updated. This caused the webmodules to be created with
the wrong state, resulting in a crash in devel within
PageVisibilityState (and probably all sorts of other badness in QA and
Gold). Now, the logic in StartOrResumeInternal() has been split out into
StartOrResumeInternalPreStateUpdate() and
StartOrResumeInternalPostStateUpdate(), and the navigation does not
occur until the post state update step, ensuring that the app state is
correct when it happens.

Additionally, navigations are now blocked from occurring when Cobalt is
in a suspended state. Instead, they now simply set
|pending_navigate_url_| and the navigation occurs when Cobalt resumes.

This fixes b/65600548.

Change-Id: I398182f6539a06678ce4c51259db236b63efae3c

Former-commit-id: 67e2129a35880d15ea78e4bc4acb6bdbbd6ed19d