xb1: eglGetDisplay enables PATH_FAST_ANGLE
ANGLE has a mode that removes a buffer copy at the cost of some
(unnecessary for Cobalt) compliance.

Here we're setting up to always configure this mode on, because
making the change in common Cobalt code is nonideal.

Combined with shader preloading, this also seems to eliminate
b/64877037

For cherry-pick to rc_12

Change-Id: I8d75f9c34138fee167ad27612274dfae810feef4

Former-commit-id: dffc3c47ee05f82b4e6052ac1f0336444b84c3f1