Clean up default console values
With the fps_overlay, it no longer makes sense to include rasterizer
times in the default console values. It just clogs up the console
with values that are easier to view somewhere else.

Also, added a number of values that are useful to easily have access
to, such as requested images, html element counts, box counts, and
video start delay.

Change-Id: I23eb20c79276ae392498cd7b2be3f052fcab4f71

Former-commit-id: c21479eb81865306a55eebef4e0a2f45ba97dc83