Add SOP protection to CSS Style Sheet and Script.
Moved Origin from dom/URLUtils to loader/.
When fetched cross-origin, a CSS Style Sheet should not be accessible to
the website javascript and a script should mute its error report.
This behavior is due to the reason that Cobalt does not read
crossorigin attribute of dom elements and can not know if the resource author
wants to allow CORS. So every cross-origin resource has to be protected
from the website javascript.

b/65131843
b/65132093

Change-Id: I7d49e709cf68924bb095c22d2cd8300c3debd50d

Former-commit-id: e5417efdbc1d022908783c67bcd57e60061ed7dd