Adds support for forwarding.
In some cases, such as running webdriver benchmarks on Android,
some extra work may need to be done to access services on the target
device, such as forwarding messages, etc.  This CL adds support for
this to the new launcher system.  None of the changes here will actually
be running immediately because of the old launcher system handles this
logic for the time being.  However, it still needs to be here to support
the required use cases after the old launcher system is deprecated.

Change-Id: I74c8858543d842f7ccef15200f60e12bf7a112e0

Former-commit-id: e57424374d6b1bbbe4ef72a0d3c737271be50115