Improve webdriver benchmark ad logic
Previously, the BrowseToWatch test would check for the skippable ad
button. If it existed, it knew it was in an ad and would wait for it
to become enabled. However, this failed in the case where the skippable
ad button took a while to appear and when the ad was not skippable.

The logic has been changed to check for 'ad-showing'. This appears
sooner than the skippable ad button and appears for both skippable
and unskippable ads. In the case where the ad is not skippable, the
logic now waits for the ad to finish before continuing on.

Change-Id: I1d99269adb8afb2920ff9bb0d5668072bb407472

Former-commit-id: 1b5f9e10d6d5c6628b62e5d2b4a675d6c4503f86