Drop 'win-lib'.
In theory, this is the UWP-version of win-win32-lib, however, it's missing
most things, isn't tested at all, and probably won't be used for awhile
(and when it is, I suspect we'll just want to start from scratch).

This way there's less updating needed and no confusion with win-win32-lib.

Change-Id: I1249d8fbf8b4c9621f353a6aa0b93f9ce4d1a921

Former-commit-id: cd21371cc20dd9756c9d0688368b5342e5f65ccd