Fix ANGLE GLES3 support.
eglGetDisplay now uses 'PresentFastPath'.

Unfortunately, as per Renderer11.cpp, 'PresentPathFast may not be
conformant which prohibits EGL_OPENGL_ES3_BIT_KHR being added to
the conformant flags when generating valid EGL configs even when
using D3D_FEATURE_LEVEL_11_0.
(https://cs.corp.google.com/cobalt/third_party/angle/src/libANGLE/renderer/d3d/d3d11/Renderer11.cpp?l=1067)

This means that even when specifying that we want a GLES3 context
when creating the config, it creates a GLES2 context instead (which
is somewhat surpising, I'd expect it to complain that we have no
supported configs instead rather than ignoring our option..)

When GLES3_SUPPORTED is defined (as it is when we're using
FEATURE_LEVEL_11), we expect GLES3 support and the fact that we're
not using GLES3 is likely causing issues in win-win32 as well as
causing a crash for YouTube VR with win-win32-lib.

Change-Id: Ib318a5b7a80c59a4381dc3654dafb9a571c8c07d

Former-commit-id: 0123d5df3a68515c5cb1d95539957dbea8986a2f