Schedule resize when root element fails to provide offsets
When Cobalt is in a suspended state, layouts are disabled. This can
cause issues when Cobalt had the load event occur right before the
suspend. In this case, the root element is unable to provide offset
dimensions, which causes ***REMOVED*** to set the root element's font-size
to 0%, resulting in most of the UI being invisible.

To account for this, the root element now notifies the document when
it has its offset dimensions requested when it is unable to provide
them. This causes a resize to be scheduled for when the visibility
changes and layouts are again available.

This fixes b/65854828.

Change-Id: I5c1738ee7130c4c0a10236e119ba88a3d8296141

Former-commit-id: 8ecf77d8312ca8b93c430ed7bacef3b08f705b4c