xb1: Fix includes in windows builds.
Previously, a header that was being force included on linux was missing
in the Visual Studio builds.  This is due to the fact there are
different gyp variables that need to be modified to force include files.

This CL brings the windows and linux build closer together by using the
correct gyp variables on windows.

b/65687585

Change-Id: I7fd15a4fdf45304dfbf942c4889db18b6f99b9d7

Former-commit-id: a8cea28b2b68d021fdeb2b83aa807469367b3781