Make BrowseToWatch webdriver benchmark work with Airstream
The BrowseToWatch test did not work with Airstream because escape no
longer causes the focus to jump up to the top shelf. Instead, it
opens up the guide. To account for this, the logic has changed so that
videos are played until the action tile at the bottom is reached. It
is then pressed which causes the focus to jump back up to the top.

Additionally, the logic has been improved so that channel tiles are
skipped. Previously, they'd be entered and would exit after a video
failure timeout.

The logic checks for both the airstream and non-airstream versions of
the tiles.

Change-Id: Ia843bf0d0d41ecf25c14f8f350ee86ca6bce1453

Former-commit-id: e8b0fa739a391b881baffb6d7fe6552db39cce07