Don't crash in QA and Gold when there is no application
Previously, we'd DCHECK in Debug and Devel but crash in QA and Gold
when we received a starboard event without an existing application.

We now log when this occurs, still DCHECK in Debug and Devel, but
return without crashing in QA and Gold.

Additionally, PreloadApplication(), StartApplication(), and
StopApplication() events are now all logged.

Change-Id: I37dbad512018583a6378a3901b85199a0af4183e

Former-commit-id: ff638f7aee5d73aa739d9e03060b8b627d2d9c8f