Add undefined behavior and unsafe math nplb tests
This change adds two nplb tests, undefined behavior and unsafe math.
The undefined behavior test tests for an optimizing compiler making
the dangerous but legal optimization of removing null this pointer
checks.  The unsafe math test tests the correctness of several
floating point edge cases, such as operations with NaN values,
infinity, and signed zero.

Also, the -fno-delete-null-pointer-checks is added to
linux-x64x11-gcc-6-3, in order for it to pass the undefined behavior
test.

Change-Id: Icdf70b5bbee72aeca6d370810701fc45e2d9795c

Former-commit-id: 83ae6500bf955464e3d0b0be57d931f3ca1a257f