Add tracking of loaded resources and total loaded memory
Previously, the resource caches tracked the requested resources, but not
the successfully loaded total. This caused issues with image counts
because non-images with no underlying data are requested through the
image cache. Now, we track images with underlying data that were
successfully loaded.

Additionally, we now track the total loaded memory, rather than just
the current memory in the cache. This allows us to easily track the
total memory being loaded by each specific event without needing to
worry about the amount that the cache is purging.

Also, SkTypeface_cobalt previously used an estimate on how large the
typeface's data size is. This has been improved to use the size of the
underlying stream, which will provide an exact size.

These changes will allow us to track image and font loaded counts and
memory on the events that are tracked by the webdriver benchmarks.

Change-Id: I61235577d35d47bcda692be292c64ea15eb64985

Former-commit-id: 4b68f892cae028a6b704363cfe5afdb855468885