Adds new features to the launcher abstraction.
The following have been added:

- The ability to specify an output file for the launcher to
write to. This defaults to sys.stdout unless otherwise specified.

- The ability to specify the "out" directory.
This will allow us a way to move to the new launcher system while
keeping special cases such as the "source" build working.

- The ability to specify a startup timeout in seconds.
This has been carried over from the old launcher.

In addition, signal handling has been removed from the launchers
themselves.  The intention is for the clients to handle this, calling
the launcher's "kill" method to ensure a clean exit.

Change-Id: Ic780639ec8a11bd00abfc04872cbe39ad7c22175

Former-commit-id: 965b3b1e74aed9549d0035d6bdbe73306d5f3ece