Polish up splash screen transition out animations.
Introduces a TimeFence() function to renderer::Pipeline
that when called, ensures that the renderer renders up to
the specified time before processing additional submissions.
This is helpful to ensure the final frames of the splash
screen transition out animation are displayed.

Provide a timestamp from the document's timeline when
calling the window.close() callback that indicates when
window.close() was called.  This way, we have a time
value that we can provide to the renderer's new TimeFence()
function in order to ensure animations play out until the
end.

Introduces the ability for the pipeline to detect when
we switch from one timeline to another, and when this
happens, it will reset the SubmissionQueue, possibly
with different configuration settings.

Configure the splash screen timeline to never speed
up animations (which is fine since it's non-interactive
and so there is no need to reduce latency) and also
increase the submission queue size during splash screen
display, is it has been found to hit this resulting
in the beginning of the transition out animation being
skipped.

This gets the raspi-1 and raspi-0 displaying the splash
screen nicely.

b/65858230
b/65776065

Change-Id: I9b0a8834fa8d5be1b31d1b308ac2d9a408f82195

Former-commit-id: 2eec8c07c56a8fc466ceb0e996a2bb194a294cb0